api/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
h11==0.13.0
idna==3.3
Markdown==3.4.1
outcome==1.2.0
PySocks==1.7.1
pytz==2022.2.1
requests==2.28.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wsproto==1.1.0
